refactor: deletei o  projeto Apllication e adicionei um novo csproj
Removi as dto's da API e transferi para o projeto Application

+++ AgiliFood.Business/Models/Product.cs
@@ -16,7 +16,7 @@
         
         public string Image { get; set; }
 
-        public long ProductCategoryId { get; set; }
+        public int ProductCategoryId { get; set; }
 
         public ProductCategory ProductCategory { get; set; }
